requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
colorlog==6.10.1
contourpy==1.3.3
cycler==0.12.1
fastapi==0.129.0
filelock==3.20.0
fonttools==4.61.1
fsspec==2025.12.0
greenlet==3.3.2
h11==0.16.0
idna==3.11
imbalanced-learn==0.14.1
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
lightgbm==4.6.0
llvmlite==0.46.0
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numba==0.64.0
numpy==2.4.1
optuna==4.7.0
packaging==26.0
pandas==3.0.0
pillow==12.1.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
setuptools==70.2.0
shap==0.50.0
six==1.17.0
sklearn-compat==0.1.5
slicer==0.0.8
SQLAlchemy==2.0.46
starlette==0.52.1
sympy==1.13.1
threadpoolctl==3.6.0
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchvision==0.21.0+cu124
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
xgboost==3.1.3